+++ CHANGELOG.md
@@ -241,7 +241,7 @@ https://github.com/MAIF/otoroshi/releases/tag/v1.5.0
 - Move manual (#986)
 - New dev mode (#987)
 - PEM bundle import endpoint (#988)
-- Add a log4jshell mitigation plugin (#989)
+- Add a log4shell mitigation plugin (#989)
 
 ## [1.5.0-rc.4] - 2021-12-17
 
@@ -249,7 +249,7 @@ https://github.com/MAIF/otoroshi/issues?q=is%3Aissue+label%3A1.5.0-rc.4+is%3Aclo
 https://github.com/MAIF/otoroshi/compare/v1.5.0-rc.3...v1.5.0-rc.4
 https://github.com/MAIF/otoroshi/releases/tag/v1.5.0-rc.4
 
-- Add a log4jshell mitigation plugin (#989)
+- Add a log4shell mitigation plugin (#989)
 - fix typos in documentation
 - add helm package publication in the release process